add orderCancle test
