+++ 2.0/order/index.js
@@ -55,36 +55,4 @@ export default class Order {
   async submit(){
     return apis.submitOrder(JSON.stringify(this.order))
   }
-  // async cancel(amount, privateKey){
-  //   this.sign()
-  //   const rawTx = {...this.order}
-  //   const tx = Transaction(rawTx)
-  //   tx.setData({
-  //     method:'cancelOrder',
-  //     order:this.order
-  //   })
-  //   await tx.setNonce()
-  //   await tx.sign(privateKey)
-  //   await tx.send()
-  // }
 }
-
-// this.cancel = (amount, privateKey) =>
-//     {
-//         if (!r || !v || !s)
-//         {
-//             this.sign(privateKey);
-//         }
-
-//         const order = {
-//             addresses: [owner, tokenS, tokenB],
-//             orderValues: [amountS, amountB, timestamp, ttl, salt, lrcFee, amount],
-//             buyNoMoreThanAmountB,
-//             marginSplitPercentage,
-//             v,
-//             r,
-//             s
-//         };
-
-//         return signer.generateCancelOrderData(order);
-//     };
